portal/src/main/java/com/kep/portal/config/ObjectMapperConfig.java
package com.kep.portal.config;

import com.fasterxml.jackson.annotation.JsonInclude;
import com.fasterxml.jackson.databind.DeserializationFeature;
import com.fasterxml.jackson.databind.ObjectMapper;
import com.fasterxml.jackson.databind.PropertyNamingStrategies;
import com.fasterxml.jackson.databind.SerializationFeature;
import com.fasterxml.jackson.datatype.hibernate5.Hibernate5Module;
import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
import com.kep.core.config.ZonedDateTimeDeserializer;
import com.kep.core.config.ZonedDateTimeSerializer;
import org.springframework.context.annotation.Bean;
import org.springframework.context.annotation.Configuration;
import org.springframework.context.annotation.Primary;
import org.springframework.http.converter.json.Jackson2ObjectMapperBuilder;

import java.time.ZonedDateTime;

/**
 * 직렬화, 역직렬화 포맷
 */
@Configuration
public class ObjectMapperConfig {

	@Bean
	@Primary
	public ObjectMapper objectMapper() {

		ObjectMapper objectMapper = new ObjectMapper();

		JavaTimeModule javaTimeModule = new JavaTimeModule();
		javaTimeModule.addSerializer(ZonedDateTime.class, new ZonedDateTimeSerializer());
		javaTimeModule.addDeserializer(ZonedDateTime.class, new ZonedDateTimeDeserializer());

		objectMapper
				.configure(SerializationFeature.FAIL_ON_EMPTY_BEANS, false)
				.configure(SerializationFeature.WRITE_DATES_WITH_ZONE_ID, false)
				.setSerializationInclusion(JsonInclude.Include.NON_EMPTY)
				.configure(SerializationFeature.INDENT_OUTPUT, true)

				.configure(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES, false)
				.setPropertyNamingStrategy(PropertyNamingStrategies.SNAKE_CASE)

				.registerModule(javaTimeModule)
				.registerModule(new Hibernate5Module())
		;

		return objectMapper;
	}

	@Bean
	public Jackson2ObjectMapperBuilder jacksonBuilder(ObjectMapper objectMapper) {

		Jackson2ObjectMapperBuilder builder = new Jackson2ObjectMapperBuilder();
		builder.configure(objectMapper);

		return builder;
	}
}
